refactor(docker-warm): extract BollardDockerWarmHost into host.rs (module-size ratchet)
serving.rs grew to 624 LOC (> 600 god-object ceiling) after the review
fixes. Split the production bollard host (seed/claim/reap I/O) into
docker_warm/host.rs, mirroring firecracker_warm/host.rs — engine policy
stays in serving.rs (444 LOC), concrete Docker I/O in host.rs (183 LOC).
Pure move; no behavior change. Also refreshes two serving.rs doc
comments left stale by the structural reconcile guard.
